services/telegram-bot/requirements.txt
python-telegram-bot==22.*
httpx==0.28.*
